diamond-backend/requirements.txt
Django==5.2.10
djangorestframework
django-cors-headers
django-filter
python-dotenv
psycopg2-binary
gunicorn
google-genai
langchain
langchain-google-genai
requests